--- pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/vedlegg/OpplysningerBruktIBeregningUT/MainTable.kt
@@ -1,2 +1,0 @@
-package no.nav.pensjon.brev.maler.vedlegg.OpplysningerBruktIBeregningUT
-
