+++ pensjon-brevbaker/src/main/kotlin/no/nav/pensjon/brev/maler/vedlegg/opplysningerbruktiberegningufoere/MainTable.kt
@@ -1,0 +1,2 @@
+package no.nav.pensjon.brev.maler.vedlegg.opplysningerbruktiberegningufoere
+
